==> 4_table1.Rmd <==
---
title: "Generating Table 1"
output: 
  html_document:
    toc: true
    keep_md: true
---


```{r,message=FALSE}
library(tidyverse)
library(survey)
library(plyr)
library(dplyr)
library(factoextra)
library(labelled)
library(tableone)
library(memisc)
library(jtools)
library(Publish)
```

# Setting up working directory
```{r}
base_path <- "~/Dropbox (Partners HealthCare)/SOL_misc_genetics/survey_gen_test_utilization/20221003_data_code/"
data_file <- paste0(base_path, "Processed_data/gte_data_set_with_covariates_and_IPW.RData")
output_dir <- paste0(base_path, "Results")
```

# Loading the data and weight the data with IPW (trim the weights at 30)
```{r}
data_gte <- readRDS(data_file)
nrow(data_gte)


survey_gte <- svydesign(id=~PSU_ID, strata=~STRAT, weights=~IPW , data=data_gte)
survey_trim <- trimWeights(survey_gte,upper = 30)
```


# Generate the table without stratification
```{r}
tbl1_var_1<-c("AGE","SEX","Education","Income_level","Current_Health_insurance","Physician_Visit", "Background", "Language_pref", "Marital_status","Employment_status", "US_BORN")
tb1_noweight<- print(CreateTableOne(vars = tbl1_var_1,data =data_gte ),missing=TRUE,varLabels = TRUE,digits =3,pDigits=3,showAllLevels=TRUE)

tbl1_weighted<- print(svyCreateTableOne(vars = tbl1_var_1,  data =survey_trim), missing=TRUE, varLabels = TRUE,digits =3,pDigits=3, showAllLevels=TRUE)
```
Combine the count values (from the unweighted table) and percentage (from the weighted table) together

```{r}
tbl1_w <- tbl1_weighted 
tbl1 <- tb1_noweight

tbl1_comb <- tbl1
# ignore the "Miami" typo...
col_ind_to_update <- which(colnames(tbl1_w) %in% c("Overall"))

# update tbl1_comb with the percentages from the weighted table
for (i in col_ind_to_update){
  counts <- sapply(tbl1[,i], function(x){
                  strsplit(x, split = "(", fixed = TRUE)[[1]][1]
                 })
  percnt <- sapply(tbl1_w[,i], function(x){
                paste0("(",  strsplit(x, split = "(", fixed = TRUE)[[1]][2])
              })
  tbl1_comb[,i] <- paste0(counts, percnt)
}

  
write.csv(tbl1_comb,paste0(output_dir,"/tbl1_unstratified.csv"))

```

# Generate a table stratified by study center
The table follows the same pattern as the unstratified table.

```{r}
tb1_noweight_center <- print(CreateTableOne(vars = tbl1_var_1,strata = "CENTER",data = data_gte),missing=TRUE,varLabels = TRUE,digits =3,pDigits=3,showAllLevels=TRUE)
tbl1_weighted_center<- print(svyCreateTableOne(vars = tbl1_var_1, strata = "CENTER", data = survey_trim), missing=TRUE, varLabels = TRUE,digits =3,pDigits=3,showAllLevels=TRUE)

```

Try the same thing without turning the table into a data frame, to have nicer row names:
```{r}
tbl1_wc <- tbl1_weighted_center[,-which(colnames(tb1_noweight_center) %in% c("p", "test"))]   
tbl1_c <- tb1_noweight_center[,-which(colnames(tb1_noweight_center) %in% c("p", "test"))]

tbl1_comb <- tbl1_c
# ignore the "Miami" typo...
col_inds_to_update <- which(colnames(tbl1_wc) %in% c("Bronx", "Chicago", "Miami", "San Diego"))

# update tbl1_comb with the percentages from the weighted table
for (i in col_inds_to_update){
  counts <- sapply(tbl1_c[,i], function(x){
                  strsplit(x, split = "(", fixed = TRUE)[[1]][1]
                 })
  percnt <- sapply(tbl1_wc[,i], function(x){
                paste0("(",  strsplit(x, split = "(", fixed = TRUE)[[1]][2])
              })
  tbl1_comb[,i] <- paste0(counts, percnt)
}

  
write.csv(tbl1_comb,paste0(output_dir,"/tbl1_stratified.csv"))
```


# version:
```{r}
sessionInfo()
```
