==> 8_bar_plots_unstratified.Rmd <==
---
title: "Generate bar plots based on responses to genetic testing questions"
output: 
  html_document:
    toc: true
    keep_md: true
---

```{r,message=FALSE}
library(tidyverse)
library(survey)
library(plyr)
library(dplyr)
library(factoextra)
library(labelled)
library(tableone)
library(memisc)
library(jtools)
library(Publish)
```

# Setting up working directory
```{r}
base_path <- "~/Dropbox (Partners HealthCare)/SOL_misc_genetics/survey_gen_test_utilization/20221003_data_code/"
data_file <- paste0(base_path, "Processed_data/gte_data_set_with_covariates_and_IPW.RData")
output_dir <- paste0(base_path, "Results")
```



# read the data and format variables as necessary

```{r}
data_gte <- readRDS(data_file)


# change "refuse to respond" to the individual awareness items to 0:
data_gte$GTE1[which(data_gte$GTE1 == 9)] <- 0
data_gte$GTE2[which(data_gte$GTE2 == 9)] <- 0
data_gte$GTE3[which(data_gte$GTE3 == 9)] <- 0
data_gte$GTE4[which(data_gte$GTE4 == 9)] <- 0

# change anything that is not 1 or 0 in response to "offered" and "used" questions to 0:
data_gte$GTE1A[which(!is.element(data_gte$GTE1A, c(0,1)))] <- 0
data_gte$GTE2A[which(!is.element(data_gte$GTE2A, c(0,1)))] <- 0
data_gte$GTE3A[which(!is.element(data_gte$GTE3A, c(0,1)))] <- 0
data_gte$GTE4A[which(!is.element(data_gte$GTE4A, c(0,1)))] <- 0

data_gte$GTE1B[which(!is.element(data_gte$GTE1B, c(0,1)))] <- 0
data_gte$GTE2B[which(!is.element(data_gte$GTE2B, c(0,1)))] <- 0
data_gte$GTE3B[which(!is.element(data_gte$GTE3B, c(0,1)))] <- 0
data_gte$GTE4B[which(!is.element(data_gte$GTE4B, c(0,1)))] <- 0

# add composite responses:

# aware:
data_gte$aware <- NA
data_gte$aware[which(data_gte$GTE1 == 1 | 
                         data_gte$GTE2 == 1 | 
                         data_gte$GTE3 == 1 | 
                         data_gte$GTE4 == 1)] <- 1
data_gte$aware[which(is.na(data_gte$aware) & 
                         (data_gte$GTE1 == 0 | 
                         data_gte$GTE2 == 0 | 
                         data_gte$GTE3 == 0 | 
                         data_gte$GTE4 == 0))] <- 0

# offered:
data_gte$offered <- 0
data_gte$offered[which(data_gte$GTE1A == "1" | 
                         data_gte$GTE2A == "1" | 
                         data_gte$GTE3A == "1" | 
                         data_gte$GTE4A == "1")] <- 1

# received:
data_gte$received <- 0
data_gte$received[which(data_gte$GTE1B == "1" | 
                         data_gte$GTE2B == "1" | 
                         data_gte$GTE3B == "1" | 
                         data_gte$GTE4B == "1")] <- 1

survey_gte <- svydesign(id=~PSU_ID, strata=~STRAT, weights=~IPW , data=data_gte)
survey_trim <- trimWeights(survey_gte,upper = 30)
```


# Compute proportions and SEs and print to file
The survey R package is a bit annoying, we cannot extract the SE of proportion estimates directly. So we write them to a file. 

First, set up the variable names and the files. We want to describe these results by gender as well, these will be in separate files. 
```{r}
vars <- c("aware", "offered","received",
          "GTE1","GTE1A","GTE1B",
          "GTE2","GTE2A","GTE2B",
          "GTE3","GTE3A","GTE3B",
          "GTE4","GTE4A","GTE4B")
gte_prop_commands_file_name <- paste0(output_dir, "/proportion_and_SEs_commands_among_", 
                                      c("all", "Females", "Males"), ".txt")
gte_prop_file_name <- paste0(output_dir, "/proportion_and_SEs_among_", 
                                      c("all", "Females", "Males"), ".txt")
names(gte_prop_commands_file_name) <- names(gte_prop_file_name) <- c("all", "f", "m")
```

Write information to files.
```{r, eval = FALSE}

# first we print the commands (otherwise we cannot efficiently loop through variables)
sink(gte_prop_commands_file_name[["all"]])
 for (cur_var in vars){
 
    line <- paste('print(svyciprop(~I(', cur_var, '==1),survey_trim,\"lo\"))')
    cat(line, "\n")
 
 }
 sink()

# now execute the commands and print results to file
sink(gte_prop_file_name[["all"]])
 source(gte_prop_commands_file_name[["all"]])
sink()

# now by gender groups
# First, women:
sink(gte_prop_commands_file_name[["f"]])
 for (cur_var in vars){
 
    line <- paste('print(svyciprop(~I(', cur_var, '==1),subset(survey_trim, SEX == \"Female\"),\"lo\"))')
    cat(line, "\n")
 
 }
 sink()

# now execute the commands and print results to file
sink(gte_prop_file_name[["f"]])
 source(gte_prop_commands_file_name[["f"]])
sink()

## In men:
sink(gte_prop_commands_file_name[["m"]])
 for (cur_var in vars){
 
    line <- paste('print(svyciprop(~I(', cur_var, '==1),subset(survey_trim, SEX == \"Male\"),\"lo\"))')
    cat(line, "\n")
 
 }
 sink()

# now execute the commands and print results to file
sink(gte_prop_file_name[["m"]])
 source(gte_prop_commands_file_name[["m"]])
sink()
```

Read and organize the data:
```{r}
# read the results from the file, first: everyone combined. 
bar_dat_list <- vector(mode = "list", length = 3)
groups <- c("all" , "f", "m")
names(bar_dat_list) <- groups

for (i in 1:3){
  group <- groups[i]
  bar_dat <- c()

  for (rownum in seq(from =2, to = length(vars)*2, by = 2)){
  bar_dat <- rbind(bar_dat, read.table(gte_prop_file_name[[group]], header = FALSE, 
                                       nrow =1, skip = rownum-1))
  }

  # organize
  bar_dat$question <- vars
  bar_dat$V1 <- bar_dat$V2 <- bar_dat$V3 <- NULL
  names(bar_dat)[1:3] <- c("Proportion", "Low_CI", "High_CI")
  bar_dat$Question <- rep(c("Aware","Offered","Used"),times=5)
   bar_dat$Genetic_Test <- c(rep("Any test type", 3), rep("Disease Risk",3),
                             rep("Risks to Children",3),
                             rep("Personalized Treatment",3),rep("Drug Efficacy",3))
  
  bar_dat$group <- group
  bar_dat_list[[group]] <- bar_dat
}
bar_dat_comb <- do.call(rbind, bar_dat_list)
bar_dat_comb$group[which(bar_dat_comb$group == "all")] <- "All"
bar_dat_comb$group[which(bar_dat_comb$group == "m")] <- "Men"
bar_dat_comb$group[which(bar_dat_comb$group == "f")] <- "Women"
```


# Generate the bar plot: all combined
 The x-axis is the 4 questions (and composite response) in the test and the y-axis is the proportions of response 'Yes'.
 There are 5 bars in each group (question) and they represent the 4 tests and the composite measure of any of the tests. 

```{r}
bar_dat_comb$Genetic_Test <- factor(bar_dat_comb$Genetic_Test,levels =c("Any test type", "Disease Risk","Risks to Children","Personalized Treatment","Drug Efficacy"), ordered = TRUE) 
p <- ggplot(bar_dat_comb[which(bar_dat_comb$group == "All"),], aes(x=Question, y=Proportion, fill=Genetic_Test)) + 
      geom_bar(stat="identity", position=position_dodge(0.9))+ 
      geom_errorbar(aes(ymin=Low_CI, ymax=High_CI), width=.2, position=position_dodge(.9)) +
      scale_fill_manual(values=c("#fec44f", "#2b8cbe","#a6bddb", "#2ca25f", "#99d8c9")) +
      theme_minimal() + xlab("")
p
ggsave(paste0(output_dir, "/proportion_of_all_pop_bars.pdf"),width=6,height = 4)

```


# Generate the bar plot: all combined, and by gender 
Another attempt at a figure with facets for each of the analyses (combined, males, females)

```{r}

p <- ggplot(bar_dat_comb, aes(x=Question, y=Proportion, fill=Genetic_Test)) + 
      geom_bar(stat="identity", position=position_dodge(0.9))+ 
      geom_errorbar(aes(ymin=Low_CI, ymax=High_CI), width=.2, position=position_dodge(.9)) +
      scale_fill_manual(values=c("#fec44f", "#2b8cbe","#a6bddb", "#2ca25f", "#99d8c9")) +
      theme_minimal() + xlab("")
p + facet_wrap(~group)
ggsave(p + facet_wrap(~group), file = paste0(output_dir, "/proportion_strat_pop_bars.pdf"),width=10,height = 4)

```



# Create another barplot: among individuals who are aware of the test.

```{r}
vars <- c("offered", "received", "GTE1A","GTE1B",  "GTE2A","GTE2B", "GTE3A","GTE3B", "GTE4A","GTE4B")
using_var <- c(rep("aware", 2), rep("GTE1",2), rep("GTE2",2), rep("GTE3",2), rep("GTE4",2))
gte_prop_commands_file_name <- paste0(output_dir, "/proportion_and_SEs_commands_among_aware.txt")
gte_prop_file_name <- paste0(output_dir, "/proportion_and_SEs_among_aware.txt")
```

Write information to files.
```{r, eval = FALSE}

# first we print the commands (otherwise we cannot efficiently loop through variables)
sink(gte_prop_commands_file_name)
 for (i in 1:length(vars)){
  
    line <- paste('print(svyciprop(~I(', vars[i], '==1),', 
                  'subset(survey_trim,', using_var[i], '==1 ),\"lo\"))')
    cat(line, "\n")
 }
 sink()

# now execute the commands and print results to file
sink(gte_prop_file_name)
 source(gte_prop_commands_file_name)
sink()

```


Read and organize the data:
```{r}
# read the results from the file 
bar_dat <- c()
for (rownum in seq(from =2, to = length(vars)*2, by = 2)){
  bar_dat <- rbind(bar_dat, read.table(gte_prop_file_name, header = FALSE, nrow =1, skip = rownum-1))
}

# organize
 bar_dat$question <- vars
 bar_dat$V1 <- bar_dat$V2 <- bar_dat$V3 <- NULL
 names(bar_dat)[1:3] <- c("Proportion", "Low_CI", "High_CI")
 bar_dat$Question <- rep(c("Offered","Used"),times=5)
 bar_dat$Genetic_Test <- c(rep("Any test type", 2), rep("Disease Risk",2),rep("Risks to Children",2),rep("Personalized Treatment",2),rep("Drug Efficacy",2))

```


# Generate the bar plot
 The x-axis is the 4 questions in the test, and the composite response, and the y-axis is the proportions of response 'Yes'.
 There are 5 bars in each group (question) and they represent 4 tests and composite response. 

```{r}
bar_dat$Genetic_Test <- factor(bar_dat$Genetic_Test,levels =c("Any test type", "Disease Risk","Risks to Children","Personalized Treatment","Drug Efficacy"))

p <- ggplot(bar_dat, aes(x=Question, y=Proportion, fill=Genetic_Test)) + 
      geom_bar(stat="identity", position=position_dodge(0.9))+ 
      geom_errorbar(aes(ymin=Low_CI, ymax=High_CI), width=.2, position=position_dodge(.9)) +
    scale_fill_manual(values=c("#fec44f", "#2b8cbe","#a6bddb", "#2ca25f", "#99d8c9")) + 
    theme_minimal() + xlab("")
p
ggsave(p, file = paste0(output_dir, "/proportion_among_aware_pop_bars.pdf"),width=6,height = 4)

```


```{r}
sessionInfo()
```
