==> 10_compare_responders_non_responders.Rmd <==
---
title: "A supplemental table comparing responders and non-responders"
output: 
  html_document:
    toc: true
    keep_md: true
---


```{r,message=FALSE}
library(tidyverse)
library(survey)
library(plyr)
library(dplyr)
library(factoextra)
library(labelled)
library(tableone)
library(memisc)
library(jtools)
library(Publish)
library(flextable)
```

# Motivation
Many individuals did not respond to the genetic testing survey. Are these individuals different than those who responded? in what way? we create a simple table that compares the characteristics of respondents to non-respondents. To compute weights, we use probabilities of participation in the AFU. 

# Setting up working directory
```{r}
base_path <- "~/Dropbox (Partners HealthCare)/SOL_misc_genetics/survey_gen_test_utilization/20221003_data_code/"

data_file_w_afu_probs <- paste0(base_path, "Processed_data/organized_visit1_covariates_with_prob_AFU.RData")
data_file_w_visit2_covars <- paste0(base_path, "Processed_data/organized_covariates_updated_from_vis2.RData")
surve_gen_file <- paste0(base_path, "/Preprocessed_data/solfu_gte_20211122.csv")
output_dir <- paste0(base_path, "Results")
```

# Loading the data and compute weights
The weights here are based on participation in the AFU, not in the genetic testing survey.
```{r}
# load the data set with the compute probabilites of participation in the AFU
data_afu <- readRDS(data_file_w_afu_probs)
nrow(data_afu)

# compute IPW for this analysis
data_afu$IPW <- data_afu$WEIGHT_FINAL_NORM_OVERALL*
                          (1/data_afu$prob_AFU)
data_afu <- data_afu[which(data_afu$AFU_par == 1),]  
hist(data_afu$IPW)


# load the dataset with covariates with information from visit 2 when possible (rather than from visit 1)
data_2 <- readRDS(data_file_w_visit2_covars)

# create GTE participation variable
# read the genetic testing survey data: 
survey_gen <- read.csv(surve_gen_file)

# remove rows with data from afu_year =2 
survey_gen <- survey_gen[-which(survey_gen$AFU_YEAR==2),]

# create a variable that indicates whether a participate responded to the genetic testing survey according to the response to the awareness questions
survey_gen$aware <- NA
survey_gen$aware[which(survey_gen$GTE1 == 1 | 
                         survey_gen$GTE2 == 1 | 
                         survey_gen$GTE3 == 1 | 
                         survey_gen$GTE4 == 1)] <- 1
survey_gen$aware[which(is.na(survey_gen$aware) & 
                         (survey_gen$GTE1 == 0 | 
                         survey_gen$GTE2 == 0 | 
                         survey_gen$GTE3 == 0 | 
                         survey_gen$GTE4 == 0))] <- 0

survey_gen$gte_par <- NA
survey_gen$gte_par[which(survey_gen$aware == 0 | survey_gen$aware == 1)] <- 1 

gte_par_id <- unique(survey_gen$ID[which(survey_gen$gte_par == 1)])

data_2$GTE_par <- "no"
data_2$GTE_par[match(gte_par_id, data_2$ID)] <- "yes"

# Add the computed IPW to the data
data <- merge(data_2, data_afu[,c("ID", "IPW")], by = "ID")


survey_afu <- svydesign(id=~PSU_ID, strata=~STRAT, weights=~IPW , data=data)
# trim weights at 15
survey_trim <- trimWeights(survey_afu,upper = 15)
```



# Generate a table stratified by participation in the genetic testing survey

```{r}
tbl_var <-c("AGE","SEX", "CENTER","Education","Income_level","Current_Health_insurance","Physician_Visit", "Background", "Language_pref", "Marital_status","Employment_status", "US_BORN")

tbl1_noweight_gte <- print(CreateTableOne(vars = tbl_var,strata = "GTE_par",data = data),missing=TRUE,varLabels = TRUE,digits =3,pDigits=3,showAllLevels=TRUE)
tbl1_weighted_gte <- print(svyCreateTableOne(vars = tbl_var, strata = "GTE_par", data = survey_trim), missing=TRUE, varLabels = TRUE,digits =3,pDigits=3,showAllLevels=TRUE)

```

Merge weighted and unweighted tables
```{r}

tbl1_comb <- tbl1_noweight_gte
col_inds_to_update <- which(colnames(tbl1_noweight_gte) %in% c("no", "yes"))

# update tbl1_comb with the percentages from the weighted table
for (i in col_inds_to_update){
  counts <- sapply(tbl1_noweight_gte[,i], function(x){
                  strsplit(x, split = "(", fixed = TRUE)[[1]][1]
                 })
  percnt <- sapply(tbl1_weighted_gte[,i], function(x){
                paste0("(",  strsplit(x, split = "(", fixed = TRUE)[[1]][2])
              })
  tbl1_comb[,i] <- paste0(counts, percnt)
}

tbl1_comb_flex <- flextable(as.data.frame(tbl1_comb))
  
write.csv(tbl1_comb,paste0(output_dir,"/characteristics_gte_par_no_par.csv"))
flextable::save_as_docx(tbl1_comb_flex, path = paste0(output_dir, "/characteristics_gte_par_no_par.docx"))
```


# version:
```{r}
sessionInfo()
```
