==> 11_missingness_visualization.Rmd <==
---
title: "Missingness patterns in HCHS/SOL genetic testing dataset"
output: 
  html_document:
    toc: true
    keep_md: true
---


```{r,message=FALSE}
library(UpSetR)
library(naniar)
```

# Visualization of missingness patterns in the genetic testing survey dataset

## Setting up working directory and files
```{r}
base_path <- "~/Dropbox (Partners HealthCare)/SOL_misc_genetics/survey_gen_test_utilization/20221003_data_code/"

data_file_w_afu_probs <- paste0(base_path, "Processed_data/organized_visit1_covariates_with_prob_AFU.RData")

gte_data_file <- file.path(base_path, "Processed_data", "gte_data_set_with_covariates_and_IPW.RData")
output_dir <- paste0(base_path, "Results")

# the covariates used in the analysis
vars <-c("AGE","SEX", "CENTER","Education","Income_level","Current_Health_insurance","Physician_Visit", "Background", "Language_pref", "Marital_status","Employment_status", "US_BORN")
```

## Missing data visualization: all HCHS/SOL participants
```{r}
dat <- readRDS(data_file_w_afu_probs)
gg_miss_upset(dat[,vars], nsets = length(vars))

```
## Missing data visualization: participants in the AFU
```{r}
gg_miss_upset(dat[which(dat$AFU_par == 1),vars], nsets = length(vars))


```

## Missing data visualization: participants who responded to the genetic testing questionnaire
```{r}
dat_gte <- readRDS(file.path(base_path, "Processed_data/gte_data_set_with_covariates_and_IPW.RData"))
gg_miss_upset(dat_gte[,vars], nsets = length(vars))

```
